feat: Incremental RunsStream (#165)
## Motivation

Our dbt Cloud instance has over 300,000 `runs`, even with the default
page size of 5000, this takes 60+ requests and shifts a lot of data that
does not change. We would like to be able to extract the `Runs` stream
using the `INCREMENTAL` replication method.

## Inspiration

Thank you to the maintainers of
https://github.com/MeltanoLabs/tap-messagebird, who recently implemented
a working state-based sync for the conversations stream
(https://github.com/MeltanoLabs/tap-messagebird/pull/45), without using
filters on the API query. This is also how I have approached the problem
for `tap-dbt`, as it does not allow filtering on `updated_at` or
`finished_at`, and any filtering is only provided on an equivalence
basis, there is no greater/less than filtering that might have proven
useful.
 
Another discussion link that initiated that update to tap-messagebird:
https://gitlab.com/meltano/sdk/-/issues/227

## Updated Functionality

The update allows the `Runs` stream to use `finished_at` as an
`INCREMENTAL` `replication_key`. Jobs are run, at which point their run
record's `finished_at` value is null until they finish, at which point
we do not expect any further changes to the record. By ordering the API
results in reverse on `finished_at`, the tap now walks back through the
list of runs and encounters the following records in order:

- Records with `finished_at = None`
- Records where `finished_at` is greater than the state bookmark
- Records that have already been synced 

The tap stops syncing once the `finished_at` value is less than the
bookmark.

## Code Changes

- Added a new class called `AccountBasedIncrementalStream` inheriting
from `AccountBasedStream` but overriding two methods:

|Method|Override behaviour|
|----|----|
|get_url_params|If a bookmark is set - `get_starting_timestamp` has a
value, adds a new order_by parameter in the URL which is the reverse of
`replication_key` - this allows the tap to run in reverse during an
INCREMENTAL sync until it hits the bookmark when it stops syncing.|
|get_records|If running in INCREMENTAL mode, checks the replication key
value against the bookmark and stops when the replication key value is
lower than the bookmark from state. |

- Change `RunsStream` to subclass the new class and add
`replication_key`

See the README.md for a more detailed explanation of how this would work
with an example.

## Outstanding Considerations

- Some initial testing has been carried out, but I will put this into
our dev environment this week by referring to my branch in `meltano.yml`
and seeing how it performs.
- Open to discussion about how this should be structured, I'm fairly new
to subclassing and the SDK in general.
- Would appreciate some advice on if it would be better to keep
`RunsStream` a `FULL_TABLE` sync by default to make this change
non-breaking. Is it possible to do that by removing `replication_key` or
forcing `replication_method` in the `RunsStream` class? It could then be
overriden by config.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>